# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Game Realtime GM HTTP API
  version: 2.0.0
  summary: 当前单体 Go 服务的 HTTP 机器契约
  description: |
    本文件描述 `backend/internal/router/router.go` 中的 legacy 公共 HTTP 及 V2 查询接口。
    V2 路径仅在 GAMEPLAY_MODE=v2 中注册，内部事件监听器不属于公共契约。
    所有 HTTP 响应都包含 `X-Request-ID` 响应头。WebSocket `/ws` 的连接与消息契约
    由 `docs/ws-protocol.md` 单独维护，不在 OpenAPI 中建模业务消息。
...
servers:
  - url: http://localhost:8080
    description: 本地开发服务
...
paths:
...
  /api/v2/players/search:
    get:
      tags: [V2 PVE]
      operationId: searchV2Players
      summary: 搜索可交互玩家
      security: [{PlayerBearer: []}]
      parameters:
        - {name: q, in: query, required: true, schema: {type: string, maxLength: 128}}
      responses: {"200": {description: 玩家搜索结果}, "401": {$ref: "#/components/responses/Unauthorized"}}
...
  /api/v2/friends/requests:
    get:
      tags: [V2 PVE]
      operationId: listV2FriendRequests
      summary: 查询好友申请
      security: [{PlayerBearer: []}]
      responses: {"200": {description: 好友申请列表}}
  /api/v2/friends/blocks:
    get:
      tags: [V2 PVE]
      operationId: listV2Blocks
      summary: 查询拉黑对象
      security: [{PlayerBearer: []}]
      responses: {"200": {description: 拉黑列表}}
...
  /api/v2/messages/unread:
    get:
      tags: [V2 PVE]
      operationId: listV2Unread
      summary: 查询私聊未读
      security: [{PlayerBearer: []}]
      responses: {"200": {description: 未读统计}}
  /api/v2/me/activity:
    get:
      tags: [V2 PVE]
      operationId: getV2ActivitySnapshot
      summary: 查询本人统一活动恢复视图
      security: [{PlayerBearer: []}]
      responses: {"200": {description: Party/票据/候选/Run/逐人结果和待处理操作的组合视图}}
...
  /api/v2/recruitment/applications:
    get:
      tags: [V2 PVE]
      operationId: listV2RecruitmentApplications
      summary: 查询本人或自己招募帖的申请
      security: [{PlayerBearer: []}]
      responses: {"200": {description: 已脱敏的申请状态列表}}
  /api/v2/operations/catalog:
    get:
      tags: [V2 PVE]
      operationId: listV2OperationCatalog
      summary: 查询已发布关卡规则目录
      security: [{PlayerBearer: []}]
      responses: {"200": {description: 关卡、规则版本、共同目标和标签}}
  /api/v2/operations/preview:
    get:
      tags: [V2 PVE]
      operationId: previewV2TaskCompatibility
      summary: 预览个人任务兼容性
      security: [{PlayerBearer: []}]
      parameters:
        - {name: rule_version, in: query, required: true, schema: {type: string}}
        - {name: task_key, in: query, required: false, schema: {type: string}}
      responses: {"200": {description: 关卡准入、任务兼容原因、机会和结算时点}}
  /api/v2/operations/recommendations:
    get:
      tags: [V2 PVE]
      operationId: listV2OperationRecommendations
      summary: 查询本人任务的兼容关卡推荐
      security: [{PlayerBearer: []}]
      parameters: [{name: task_key, in: query, required: false, schema: {type: string}}]
      responses: {"200": {description: 仅基于公开规则和本人选择的有限推荐}}
...
components:
  securitySchemes:
    PlayerBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 仅接受 `subject_type=player` 且 `player_id>0` 的 HS256 JWT。
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 仅接受 `subject_type=admin` 且 `admin_id>0` 的 HS256 JWT。
  headers:
    RequestId:
      description: 经校验的请求追踪 ID；非法或缺失时由服务端生成。
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[A-Za-z0-9_.-]+$"
...
  responses:
...
    Unauthorized:
      description: 登录凭据错误或缺少有效身份。
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: {type: integer}
        message: {type: string}
        data:
          description: 仅部分错误返回，例如封禁原因。
...